Add some runtime asserts
There's no type safety assuring us that `Frame::pixels::<Pixel>()`
is parameterized with the same type that OpenNI2 is actually
returning.

(I tried strict typing. It works. But you can change a stream's
video mode at will, and working around that is a nightmare.)

This adds a one-time assertion to the initial callback register,
and an assertion on every invocation of `Frame::pixels` (because
a callback can stay registered while a stream's video mode changes)
to ensure that type `Pixel` has the same byte size as the pixel
format according to OpenNI2.

A quirk of YUV color mode is that its pixels array element represents
2 pixels, not 1, so OpenNI2's native function for returning pixel
size returns 2 bytes, when the size of the array element is 4 bytes.
It's technically correct, but it breaks the assertion.