Enforce stream/frame pixel types
Previously the API required the user to call `Stream::read_frame()` with
a type parameter indicating the size of the pixels in the frame. Runtime
panics would ensure that the byte size of the requested pixels matched
the byte size of the actual pixels, but the user could request a frame
of any pixel format from any stream.

Instead, link the semantic pixel formats (previously an enum) with the
actual pixel structures (the structs imported from openni2-sys, all
implementing `Pixel`) so that the user cannot pair incompatible ones
together.

Semantic pixel formats are now structs that implement `Pixel`, whose
underlying pixel structure types are defined by an associated type,
and whose pixel byte sizes are defined by an associated const.

`Stream::read_frame()` no longer accepts type params, and a `Stream<P>`
returns frames of type `Frame<P::Format>`, eliminating the need to
assert and possibly panic at runtime.

It is now required to specify both the sensor type as an enum argument
and a pixel format as a type paramenter when creating a stream. It is
still impossible to pair invalid sensors and pixel formats together,
but that failure is encoded in a `Result` returned by
`Device::create_stream()` rather than a panic when the pixels are finally
read.